66_bug_fix

+++ plugins/lb_target.go
@@ -242,7 +242,7 @@ func ensureHostAddToLbPool(params CloudProviderParam, hostIds []string, hostPort
 		address, _ := getIpFromVmInfo(vm)
 
 		//check if already exist
-		if _, err = getMemberIdByIpAndPort(allMembers, address, hostPorts[i]);err==nil{
+		if _, err = getMemberIdByIpAndPort(allMembers, address, hostPorts[i]); err == nil {
 			continue
 		}
 
@@ -368,6 +368,8 @@ func ensureDeleteHostFromPool(params CloudProviderParam, hostIds []string, hostP
 		return err
 	}
 
+	alreadyDeletedHost:=make(map[string]bool)	
+
 	for i, hostId := range hostIds {
 		vm, err := getVmInfoById(params, hostId)
 		if err != nil {
@@ -378,11 +380,15 @@ func ensureDeleteHostFromPool(params CloudProviderParam, hostIds []string, hostP
 		if err != nil {
 			continue
 		}
+		if _,ok:=alreadyDeletedHost[memberId];ok{
+			continue
+		}
 
 		if err = pools.DeleteMember(sc, poolId, memberId).ExtractErr(); err != nil {
 			logrus.Errorf("lb pools deleteMember meet err=%v", err)
 			return err
 		}
+		alreadyDeletedHost[memberId]=true
 	}
 
 	return nil
